# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 50
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  dto.CreateTagRequest:
    properties:
      name:
        maxLength: 50
        type: string
      slug:
        type: string
    required:
    - name
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 50
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  dto.UpdateTagRequest:
    properties:
      name:
        maxLength: 50
        type: string
      slug:
        type: string
    required:
    - name
    type: object
  handler.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
    type: object
  handler.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/handler.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: eastdong1106@gmail.com
    name: API Support
  description: 개인 블로그 백엔드 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
...
  /api/admin/categories:
    get:
      description: Get a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List all categories (admin)
      tags:
      - admin/categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - admin/categories
  /api/admin/categories/{id}:
    delete:
      description: Delete a category (must have no posts)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - admin/categories
    get:
      description: Get a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a category by ID (admin)
      tags:
      - admin/categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - admin/categories
...
  /api/admin/tags:
    get:
      description: Get a list of all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List all tags (admin)
      tags:
      - admin/tags
    post:
      consumes:
      - application/json
      description: Create a new tag
      parameters:
      - description: Tag data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new tag
      tags:
      - admin/tags
  /api/admin/tags/{id}:
    delete:
      description: Delete a tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a tag
      tags:
      - admin/tags
    get:
      description: Get a single tag by its ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a tag by ID (admin)
      tags:
      - admin/tags
    put:
      consumes:
      - application/json
      description: Update an existing tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a tag
      tags:
      - admin/tags
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer 토큰을 입력하세요 (예: Bearer eyJhbGc...)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"